refactor(ui): plugin load pipeline moves behind a loader service
Tranche 7. New services/plugin_loader.rs owns the whole two-phase
plugin lifecycle: the background phase-1 loaders
(load_plugin_effect_bg / load_plugin_instrument_bg +
plugin_device_ref), the project-reload thread
(spawn_device_reloads), the result structs, and new
finish_effect_init / finish_instrument_init for phase 2 (UI-thread
init + saved-state restore + GUI/state pointer capture).

poll_plugin_loads shrinks to: receive, finish via the service,
apply to state and engine. No teardown-ordering changes anywhere.
Smoke-tested live: app runs, audio thread promotes, CLAP main
thread registers.
