+++ crates/vibez-ui/src/services/mod.rs
@@ -1,0 +1,4 @@
+//! Long-running side-effect services: background work with explicit
+//! inputs and channel outputs, kept out of the update loop.
+
+pub mod plugin_loader;
